break up Heatmap to Heatmap + GetHeatMapMonth
